feat: skip downloaded of weights if already there

+++ BIGDreamBooth/executor/executor.py
@@ -69,7 +69,7 @@ class BIGDreamBoothExecutor(Executor):
         )
         if self.models_dir:
             os.makedirs(self.models_dir, exist_ok=True)
-        download_pretrained_stable_diffusion_model(self.models_dir)
+            download_pretrained_stable_diffusion_model(self.models_dir)
 
         self.user_to_identifiers_and_class_names: Dict[str, Dict[str, str]] = defaultdict(lambda: defaultdict(str))
 
@@ -229,10 +229,10 @@ class BIGDreamBoothExecutor(Executor):
 
 def download_pretrained_stable_diffusion_model(model_dir: str, sd_version: str = 'stable-diffusion-v1-4'):
     """Downloads pretrained stable diffusion model."""
-    # todo: skip download if we can load pretrained weights from disk
-    pipe = StableDiffusionPipeline.from_pretrained(f"CompVis/{sd_version}", use_auth_token=True)
-    for dir in [PRE_TRAINDED_MODEL_DIR, METAMODEL_DIR]:
-        pipe.save_pretrained(os.path.join(model_dir, dir))
+    if not all(os.path.exists(os.path.join(model_dir, dir)) for dir in [PRE_TRAINDED_MODEL_DIR, METAMODEL_DIR]):
+        pipe = StableDiffusionPipeline.from_pretrained(f"CompVis/{sd_version}", use_auth_token=True)
+        for dir in [PRE_TRAINDED_MODEL_DIR, METAMODEL_DIR]:
+            pipe.save_pretrained(os.path.join(model_dir, dir))
 
 
 def cmd(command, std_output=False, wait=True):
